refactor(task-completion): Phase 5 cleanup - remove feature flag and legacy code
Removed feature flag and all conditional logic now that signal-based completion
is the default behavior. This completes the migration from buffer/size-based
heuristics to explicit signal-based completion detection.

Changes:
- Remove use_signal_based_completion feature flag from settings.py
- Remove conditional detector creation in orchestrator (always create)
- Make task_completion_detector required parameter (not Optional)
- Remove conditional checks in continuation_strategy.py
- Remove conditional checks in intent_classification.py
- Remove buffer/size heuristics from streaming handler (always return None)
- Fix import paths (victor.agent.task_completion, not protocols.task_completion)
- Update tests to pass task_completion_detector parameter

The system now exclusively uses signal-based completion detection with
TaskCompletionDetector, providing deterministic completion based on
explicit markers (_DONE_, _TASK_DONE_, _SUMMARY_) instead of legacy
buffer/size heuristics.
